requirements.txt
djangorestframework>=3.15.2,<3.16
django-cors-headers>=4.4.0,<4.5
requests>=2.32.3,<2.33